feat: Add SMS producer for sending SMS messages
This commit adds a new file `producer.go` to the `sms` package. The `ProduceSMS` function is implemented to produce messages to a Kafka topic for sending SMS notifications. The producer is configured with the necessary Kafka properties, such as the bootstrap servers.

The addition of this SMS producer enhances the functionality of the application by enabling the sending of SMS notifications.