refactor: Update SMS service to use SendSMS function
This commit refactors the `sms` package to use the newly added `SendSMS` function instead of the previous `SendSMS` method in the `Service` struct. The `SendSMS` function is a standalone function that sends an SMS message using an SMS gateway API call. This change improves code organization and simplifies the implementation of the SMS service.